[REFACTOR] Move settings to own definition file
