+++ pandora-common/src/account/index.ts
@@ -1,2 +1,3 @@
 export * from './account';
 export * from './accountRoles';
+export * from './settings';
